requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
python-snap7==2.0.2
pyzmq==27.1.1
pandas==2.2.2
openpyxl==3.1.2
xlrd==2.0.1
loguru==0.7.2
python-multipart==0.0.21
requests==2.32.5